requirements.txt
adafruit-ampy==1.1.0
asgiref==3.5.0
boto3==1.22.8
botocore==1.25.8
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
Django==3.2.13
django-cors-headers==3.11.0
django-multiselectfield==0.1.12
django-rest-passwordreset==1.2.1
django-ses==3.0.1
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
jmespath==1.0.0
mysql-connector-python==8.0.29
Pillow==9.0.0
protobuf==3.20.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
